FE-83 - Moved EmailBanner as it's own component, removed from page dash
