+++ src/components/emailBanner/EmailBanner.js
