+++ src/components/emailBanner/tests/EmailBanner.test.js
